Create schema_validation_rules file
This change adds a new file representing the schema validation rules.
This is part of the seperation work between what is in `common` and
what belongs in `pycommon`. We want the schema & validation rules to
live in the service but we want the validation _logic_ to live in
common, to ensure consistency across services.