Remove DAL from script
Due to the explicit nature of the config.json file that informs the
script, the abstraction of the DAL could silently differ, which would be
unsafe for our situation. Therefore I have removed the DAL from
this script and it will be entirely boto3 with consistent application to
each table.